backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pydantic>=2.0.0
apscheduler>=3.10.0
aiofiles>=24.0.0
websockets>=13.0
jinja2>=3.1.0
python-dateutil>=2.9.0
google-genai>=1.0.0
